fix unable to marshal Checklist struct
The Pos field of the Checklist structure has type int where it
should be float64. This commit fixes it.

+++ checklist.go
@@ -32,7 +32,7 @@ type ChecklistItem struct {
 	NameData struct {
 		Emoji struct{} `json:"emoji"`
 	} `json:"nameData"`
-	Pos int `json:"pos"`
+	Pos float64 `json:"pos"`
 }
 
 func (i *ChecklistItem) Delete() error {
